requirements.txt
streamlit
ultralytics
opencv-python-headless==4.7.0.72
numpy
pillow
torch==2.0.1
torchvision==0.15.2
